Update #6
+ Added engine/Camera class

In this commit the user can now create a Camera object inside any Scene class.
The camera calculates the coordinates from the screen to the coordinates of the world.
The box matrix of vertices that is drawn on the screen is now normalized.
In the shader the rendered cube is now a multiplication matrix of the projection, the view matrix and the vec4.