Update #9
The component class is an abstract one which defines one abstract method
update, which has to be implemented by any derived class
of the Component class. One the Component Class is initialized, an instance of GameObject class
is created and by default is set to null. This can and should be changed.
It also creates an empty optional inherited function start()