Update #9
GameObject class creates a new list of Component objects. It has 3 methods

getComponent - generic method that only accept the Type that is a class or subclass of Component. It takes
a Type as argument. It returns a cast to that component.

removeComponent - another generic method that goes through the component list
and if the component is present in the list it will remove it from list.

addComponent - method that inserts a component into the list of GameObject

update - method that calls update for all components in the GameObject

start - method that calls start for all components in the GameObject