Update #9
Create a list of GameObjects in the Scene.

method start - For every game object call method start

method addGameObjectToScene - Insert the given game object into the list of game objects in the scene.
If the engine is already running means it will be added but not started, so
we have to call it manually.