Update #12
We introduced two new attribute pointers, texture coordinates and texture ID. These coordinates updated to the GPU are
used to give the position where the texture is uploaded to the screen. The texture ID is to locate which sprite sheet we are using to draw upon the screen, since we
have a limited amount of sprites we can upload, we have to upload spritesheets, where we can specify from which one we can use the sprite.