Update #13
In SpriteRenderer all the variables as well as getTexture method have been moved to the Sprite class.

In the LevelEditorScene a new Sprite object is now being initialized instead of a SpriteRenderer.
We can now speicfy which sprite we want from the spritesheet by specifying an index.
The loadResources method also loads the spritesheets now.

In the RenderBatch we now check if there is enough room to append additional textures.

In the Renderer we now check if the sprite has a texture and if there is enough room in the renderbach has enough room.

In the Texture class we initialized width and height of the texture so that we can automatically calculate where to crop the sprites to be retrieved by and index.