Update #15
Implemented sprite z-indexing and alpha rendering.

In the GameObject class the class has a new variable z_index which is initiated with the constructor. This variable is then retrieved with a get method.

In Window now GL_BLEND is now enabled

In LevelEditorScene z-index is passed for the two mario sprites.