Update #15
Now we check if the z index is equal to the z index of the game object of the sprite. We also then create a new renderBatch with the new sprite's z-index. We then use the java's collections framework to sort all the batches inside the array list of RenderBatch objects according to their z-index. This was done by overriding the method compareTo. By doing this the elements with the higher z index will be moved to the front of other objects.