Update #16
Imported all ImGUI libraries,
Changed final_window variable to pointer_final_window, because c++ has pointers and java does not, it uses the long type to store a pointer to the window.

New variables have been added that are objects of ImGUI library, in the constructor an instance of ImGUI local object is being initialized.

The run method now init_Window method and the loop_Window method.

The init_Window now calls the main init method of the Window class, the initImGUI of the Window class and two other init methods.

The loop method has been renamed from loop to loop_Window

In the loop_Window class appropriate methods for ImGUI update methods are being placed, as well as a default window pointer if the main one doesnt
render, which will probably not render the ImGUI window.