Update #17
The ImGUI Engine class now hosts all logic for the creation of the imgui package ui. It handles everything from mouse input, button input,
font creation etc. The main method of the class is the update method which calls the startframe method. This method takes care of locations of objects on the screen.
Another method that is important to mention is the currentScene.sceneImgui. This is like another update method. This method will call the updateimgui method for the scene, as well as go through any object it has and in that object for any component it possesses will call the individual updateimgui of the component.