Update #17
All of the code for the ImGUI has been moved. In the init method we set another callback which is when we resize the window. We set the new width of the window as well as
the new height through a lambda expression. At the end of the init method we create an instance of
imGUI_Engine object and call the init method of that object.
In the loop method we call update on the currentScene as well as the new objects update method. This passes the currentScene on which the ImGUI_Engine class can then work on.