Update #18
+ Creation of the serialization

For the serialization I will implement the GSON library from google. This library serializes and stores class objects into a JSON format.
In the build.grade the gson dependency was mentioned.
All of the changes implemented are due to one simple reason, and that is because the GSON serializer cant take any objects created from a class
whose constructor asks for any parameters. This is why I have removed all the parameters in
all of the classes and have moved all of the logic to getters/setters.
Classes affected are:

Component
Sprite
SpriteRenderer
Texture

Also some variables have been changed to transient, meaning they will not be serialized.