Update #19
+ Added GameObjectDeserializer class

This class is a custom deserializer for the class GameObject. The GameObject class has the variables

name, Components_li, transform and z_index

These are all retrieved using the GSON library, after which they are assembled again into a GameObject object which is then returned.

Another important thing to mention is that inside of our Components_li there are Components, for which we call our custom ComponentDeserializer instead of the deafult one.