Update #19
+ Added ComponentDeserializer class

This class, similar to GameObjectClass is a custom deserializer for the Component class. In this class we have both the serializer and the deserializer.

Since the Component class is an abstract class, we have to be able to do a general storage of every subclass from this one, therefore we store the type
aka the type of the class, as well as its properties aka the fields of the subclass.