Update #19
In this abstract we define the class updateComponentImgui, which gives a base default method for every subclass.

We get the SUBCLASSE'S class type and fields. The line

Class<?> Type = field.getTyoe();

Is important. In it we guess which type the class is. Then for the simple types we can dynamically create a ImGui element that will be drawn on the screen.
The good thing is that if we want to do anything custom for any specific non primitive class in the ImGui we can.