Update #19
In the scene class is where we define the load method. The load saveOnExit method, like mentioned, is called on exit from the levelEditorScene. This
method initiates the GsonBuilder class, which simply calls the custom de/serializers for the classes of Component and GameObject. After that It just writes all of the gameObjects in the Scene to json file.

The load method does similar thing. It builds another Gson object which then tries to load the gameObject objects stored in the json file.
It then constructs gameObjects from the file and simply adds them to the scene with all of their saved properties.