Update #21
In the MouseListener class two new static methods are added, getOrthoX and Y.

This methods are meant to host both the normalized world x and y coordinates, as well as real world coordinates. We do this firstly by normalizing the x and y values.

This is done by this line
currentX = (currentX / (float) Window.getWidth()) * 2.0f - 1.0f;

We divide the x by the width of the window, multiply by 2 and subtract 1 to get the range between -1 and 1. These are normalized coordinates now. We then multiply the new normalized coordinates with the inverse and projection matrix to get the real world coordinates.
These will both be used when placing the tiles from normalized world coordinates to real world coordinates.