Update #21
We create a new window. This Window will now go through all of the sprites in the sprite sheet, and render them in the window. It will also check if the sprites in the window are going over the border of that window, which will render the sprites in a new line, otherwise they will dynamically be transferred to a new line.