Update #22
+ Added java/ECS/Prefab.java

Like states prefab is a gameObject component that is allowed to be moved across different scenes. In our particular case its used to create a new GameObject called Sprite_Object_Gen which is a generated sprite once the sprite been dragged and placed into the world.

KeyListener class method reshaping to make it more compact.