Update #26
Youtuber added from another source which cites a c++ version of the ImGUI implementation of docking.

We enable the configFlags docking. After we render the new frame and call the setupDockSpace.

This method initializes the space for the docking window, as well as set flags for the new docking window.