Update #26
Added default empty constructor which should return error if created.

Created a new constructor that binds the texture to the openGL.

Overridden the equals method. It checks if the object is of type Texture, if not null and checks for the same values of the two compared objects.

Added getFilepath method