Update #26
The Engine now wont crash if there are no objects in the json file.

Now we loop over every gameObject in the gameObjects list. It then checks if the gameObject has a component of spriteRenderer. If it does, then it creates a new SpriteRenderer which is the gameObject's component. It then sets that sprites texture to the filepath of the texture.