Update #27
+ Created new class java/editor/GameViewViewport

This class is responsible for initializing the viewport for the Game where the level is being built.

It calculates the largest size for the viewport as well as the centered position for the viewport which is used when rendering the size of the window on the display.

In the texture changed from texture_min and texture_mag filter for openGL to be linear, which means they stretch beyond border.